+++ ImportBundle.php
@@ -11,8 +11,17 @@
 namespace KimaiPlugin\ImportBundle;
 
 use App\Plugin\PluginInterface;
+use App\Validator\Constraints\TimesheetDeactivated;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class ImportBundle extends Bundle implements PluginInterface
 {
+    // some validator rules that we skip during import, because they are not relevant here
+    // do not use the class name directly, because that would raise the required kimai version
+
+    public const SKIP_VALIDATOR_CODES = [
+        'kimai-timesheet-87', 'kimai-timesheet-88', 'kimai-timesheet-89', // timesheet deactivated before 2.0.16
+        'kimai-timesheet-deactivated-activity', 'kimai-timesheet-deactivated-project', 'kimai-timesheet-deactivated-customer', // timesheet deactivated after 2.0.16
+        TimesheetDeactivated::DISABLED_ACTIVITY_ERROR, TimesheetDeactivated::DISABLED_PROJECT_ERROR, TimesheetDeactivated::DISABLED_CUSTOMER_ERROR, // timesheet deactivated current codes
+    ];
 }